# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CAFA Program Management Information System (PMIS) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /states:
    get:
      operationId: listStates
      tags: [states]
      summary: List active operational State options or the administrative registry
      parameters:
        - name: includeInactive
          in: query
          schema: { type: boolean, default: false }
          description: Administrative registry view; restricted to State-registry administrators.
        - name: officeStatus
          in: query
          schema: { type: string, enum: [present, absent, unknown] }
          description: Optional office-presence filter.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StateRecord" }
        "401":
          description: Not authenticated
    post:
      operationId: createState
      tags: [states]
      summary: Create a canonical State registry record
      description: Only State-registry administrators may create States. Manager assignment is read-only and remains owned by User Management.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StateInput" }
      responses:
        "201":
          description: State created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateRecord" }
        "401":
          description: Not authenticated
        "403":
          description: Not authorised to manage the State registry
        "409":
          description: Name or code conflicts with an existing State after normalisation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateConflictError" }
        "422":
          description: Validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateValidationError" }
  /states/{stateId}:
    get:
      operationId: getState
      tags: [states]
      summary: Get State registry detail and read-only references
      parameters:
        - { name: stateId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateProfile" }
        "401":
          description: Not authenticated
        "404":
          description: State not found
        "422":
          description: Malformed State ID
    patch:
      operationId: updateState
      tags: [states]
      summary: Edit canonical State fields without changing its ID
      description: Only name, code, and office address are editable here. Manager assignment remains read-only and owned by User Management.
      parameters:
        - { name: stateId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StateInput" }
      responses:
        "200":
          description: State updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateRecord" }
        "401":
          description: Not authenticated
        "403":
          description: Not authorised to manage the State registry
        "404":
          description: State not found
        "409":
          description: Name or code conflicts with an existing State after normalisation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateConflictError" }
        "422":
          description: Malformed State ID or validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateValidationError" }
  /states/{stateId}/lifecycle:
    patch:
      operationId: updateStateLifecycle
      tags: [states]
      summary: Confirm and update State operational or office status
      description: Only State-registry administrators may make a confirmed, audited lifecycle change. No related records are created, reassigned, or deleted.
      parameters:
        - { name: stateId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StateLifecycleInput" }
      responses:
        "200":
          description: State lifecycle updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateRecord" }
        "401": { description: Not authenticated }
        "403": { description: Not authorised to manage the State registry }
        "404": { description: State not found }
        "422":
          description: State ID or lifecycle input is invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateValidationError" }
  /states/{stateId}/snapshot:
    get:
      operationId: getStateSnapshot
      tags: [states]
      summary: Read State Program Report snapshot data for an existing State
      parameters:
        - { name: stateId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        "200":
          description: Snapshot data for the requested State
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateSnapshot" }
        "401":
          description: Not authenticated
        "404":
          description: State not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateNotFoundError" }
        "422":
          description: Malformed State ID
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StateValidationError" }
...
  /localities:
    get:
      operationId: listLocalities
      tags: [states]
      summary: List all localities (optionally filter by stateId)
      parameters:
        - { name: stateId, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LocalityWithState" }
...
components:
  schemas:
...
    StateRecord:
      type: object
      description: Canonical Sudan State master-data record. Geographic identity, operational eligibility, and office presence are separate facts; it has no delete lifecycle.
      required: [id, name, nameAr, code, operationalStatus, officeStatus, officeAddress, managerName, localitiesCount]
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 120 }
        nameAr: { type: string, maxLength: 120 }
        code: { type: string, maxLength: 24 }
        operationalStatus: { type: string, enum: [active, inactive] }
        officeStatus: { type: string, enum: [present, absent, unknown] }
        officeAddress: { type: ["string", "null"], maxLength: 500 }
        managerName: { type: ["string", "null"] }
        localitiesCount: { type: integer }
    StateInput:
      type: object
      required: [name, nameAr, code]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        nameAr: { type: string, minLength: 1, maxLength: 120 }
        code: { type: string, minLength: 1, maxLength: 24 }
        officeAddress: { type: ["string", "null"], maxLength: 500 }
    StateLifecycleInput:
      type: object
      required: [confirmed]
      properties:
        confirmed: { type: boolean, enum: [true] }
        operationalStatus: { type: string, enum: [active, inactive] }
        officeStatus: { type: string, enum: [present, absent, unknown] }
    StateValidationError:
      type: object
      required: [error]
      properties:
        error: { type: string, enum: [validation_failed, invalid_state_id] }
        fields:
          type: object
          additionalProperties: { type: string }
    StateConflictError:
      type: object
      required: [error]
      properties:
        error: { type: string, enum: [state_identity_conflict] }
    StateNotFoundError:
      type: object
      required: [error]
      properties:
        error: { type: string, enum: ["state not found"] }
    StateSnapshot:
      type: object
      required:
        [activeProjects, activeSectors, beneficiariesReached, activitiesCompleted,
         delayedActivities, openRisks, pendingApprovals]
      properties:
        activeProjects: { type: integer }
        activeSectors: { type: integer }
        beneficiariesReached: { type: integer }
        activitiesCompleted: { type: integer }
        delayedActivities: { type: integer }
        openRisks: { type: integer }
        pendingApprovals: { type: integer }
    Locality:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    StateProfile:
      type: object
      required: [id, name, nameAr, code, operationalStatus, officeStatus, officeAddress, managerName, localitiesCount, localities, projects]
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 120 }
        nameAr: { type: string, maxLength: 120 }
        code: { type: string, maxLength: 24 }
        operationalStatus: { type: string, enum: [active, inactive] }
        officeStatus: { type: string, enum: [present, absent, unknown] }
        managerName: { type: ["string", "null"] }
        officeAddress: { type: ["string", "null"] }
        localitiesCount: { type: integer }
        localities:
          type: array
          items: { $ref: "#/components/schemas/Locality" }
        projects:
          type: array
          items: { $ref: "#/components/schemas/StateReferenceProject" }
    StateReferenceProject:
      type: object
      required: [id, code, title, status]
      properties:
        id: { type: integer }
        code: { type: string }
        title: { type: string }
        status: { type: string }
        sector: { type: ["string", "null"] }
...
    LocalityWithState:
      type: object
      required: [id, name, stateId, stateName]
      properties:
        id: { type: integer }
        name: { type: string }
        stateId: { type: integer }
        stateName: { type: string }
        stateNameAr: { type: ["string", "null"], description: "Canonical Arabic State label for locale-aware display." }
...